Improve gen_kw validation feedback
