+++ tests/unit_tests/config/test_ensemble_config.py
@@ -6,11 +6,9 @@ from textwrap import dedent
 import pytest
 import xtgeo
 from ecl.summary import EclSum
-from lark import Token
 
-from ert.config import ConfigValidationError, ConfigWarning, EnsembleConfig, ErtConfig
-from ert.config.parsing import ConfigKeys, ContextString
-from ert.config.parsing.file_context_token import FileContextToken
+from ert.config import ConfigValidationError, EnsembleConfig, ErtConfig
+from ert.config.parsing import ConfigKeys
 
 
 def test_create():
@@ -121,7 +119,7 @@ def test_that_files_for_refcase_exists(existing_suffix, expected_suffix):
         )
 
 
-def test_ensemble_config_duplicate_node_names(setup_case):
+def test_ensemble_config_duplicate_node_names():
     duplicate_name = "Test_name"
     Path("MULTFLT.TXT").write_text("", encoding="utf-8")
     Path("FAULT_TEMPLATE").write_text("", encoding="utf-8")
@@ -144,105 +142,11 @@ def test_ensemble_config_duplicate_node_names(setup_case):
             ]
         ],
     }
-    error_match = f"key {duplicate_name!r} already present in ensemble config"
-
-    with pytest.raises(ConfigValidationError, match=error_match):
+    with pytest.raises(
+        ConfigValidationError,
+        match="GEN_KW and GEN_DATA contained duplicate name: Test_name",
+    ):
         EnsembleConfig.from_dict(config_dict=config_dict)
-
-
-@pytest.mark.usefixtures("use_tmpdir")
-def test_gen_kw_pred_special_suggested_removal():
-    with open("empty_file.txt", "a", encoding="utf-8") as f:
-        f.write("")
-    with open("config.ert", "a", encoding="utf-8") as f:
-        f.write(
-            "NUM_REALIZATIONS 1\n"
-            "GEN_KW PRED empty_file.txt empty_file.txt empty_file.txt\n"
-        )
-    with pytest.warns(
-        ConfigWarning,
-        match="GEN_KW PRED used to hold a special meaning and be excluded.*",
-    ) as warn_log:
-        ErtConfig.from_file("config.ert")
-    assert any("config.ert: Line 2" in str(w.message) for w in warn_log)
-
-
-def make_context_string(msg: str, filename: str) -> ContextString:
-    return ContextString.from_token(FileContextToken(Token("UNQUOTED", msg), filename))
-
-
-@pytest.mark.usefixtures("use_tmpdir")
-def test_gen_kw_config():
-    with open("template.txt", "w", encoding="utf-8") as f:
-        f.write("Hello")
-
-    with open("parameters.txt", "w", encoding="utf-8") as f:
-        f.write("KEY  UNIFORM 0 1 \n")
-
-    with open("parameters_with_comments.txt", "w", encoding="utf-8") as f:
-        f.write("KEY1  UNIFORM 0 1 -- COMMENT\n")
-        f.write("\n\n")  # Two blank lines
-        f.write("KEY2  UNIFORM 0 1\n")
-        f.write("--KEY3  \n")
-        f.write("KEY3  UNIFORM 0 1\n")
-
-    EnsembleConfig.from_dict(
-        config_dict={
-            ConfigKeys.GEN_KW: [
-                [
-                    "KEY",
-                    "template.txt",
-                    "nothing_here.txt",
-                    "parameters.txt",
-                ]
-            ],
-        }
-    )
-
-    EnsembleConfig.from_dict(
-        config_dict={
-            ConfigKeys.GEN_KW: [
-                [
-                    "KEY",
-                    "template.txt",
-                    "nothing_here.txt",
-                    "parameters_with_comments.txt",
-                ]
-            ],
-        }
-    )
-
-    with pytest.raises(
-        ConfigValidationError, match="config.ert.* No such template file"
-    ):
-        EnsembleConfig.from_dict(
-            config_dict={
-                ConfigKeys.GEN_KW: [
-                    [
-                        "KEY",
-                        make_context_string("no_template_here.txt", "config.ert"),
-                        "nothing_here.txt",
-                        "parameters.txt",
-                    ]
-                ],
-            }
-        )
-
-    with pytest.raises(
-        ConfigValidationError, match="config.ert.* No such parameter file"
-    ):
-        EnsembleConfig.from_dict(
-            config_dict={
-                ConfigKeys.GEN_KW: [
-                    [
-                        "KEY",
-                        "template.txt",
-                        "nothing_here.txt",
-                        make_context_string("no_parameter_here.txt", "config.ert"),
-                    ]
-                ],
-            }
-        )
 
 
 def test_that_empty_grid_file_raises(tmpdir):
